优化代码，尝试model.weight_params参与梯度传播

+++ MAS/model_class.py
@@ -41,8 +41,8 @@ class SharedModel(nn.Module):
         self.reg_params = {}
         # self.used_omega_weight = torch.tensor([0.3], requires_grad=True, dtype=torch.float64)
         # self.max_omega_weight = torch.tensor([0.3], requires_grad=True, dtype=torch.float64)
-        self.weight_params = {'used_omega_weight': torch.tensor(0.3, requires_grad=True, dtype=torch.float),
-                              'max_omega_weight': torch.tensor(0.7, requires_grad=True, dtype=torch.float)}
+        self.weight_params = {'used_omega_weight': torch.tensor(0.3, requires_grad=True),
+                              'max_omega_weight': torch.tensor(0.7, requires_grad=True)}
 
     def forward(self, x):
         return self.tmodel(x)
